Potential fix for code scanning alert no. 38: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger type (double) to avoid overflow. This can be achieved by casting one of the operands to double before performing the multiplication. Specifically, we can cast the result of VP8LFastSLog2(*val_prev) to double.

Change the multiplication VP8LFastSLog2(*val_prev) * streak to (double)VP8LFastSLog2(*val_prev) * streak.
This change should be made on line 443 of the file libwebp-1.0.3/src/dsp/lossless_enc.c.

+++ libwebp-1.0.3/src/dsp/lossless_enc.c
@@ -440,7 +440,7 @@ static WEBP_INLINE void GetEntropyUnrefinedHelper(
     bit_entropy->sum += (*val_prev) * streak;
     bit_entropy->nonzeros += streak;
     bit_entropy->nonzero_code = *i_prev;
-    bit_entropy->entropy -= VP8LFastSLog2(*val_prev) * streak;
+    bit_entropy->entropy -= (double)VP8LFastSLog2(*val_prev) * streak;
     if (bit_entropy->max_val < *val_prev) {
       bit_entropy->max_val = *val_prev;
     }
